py-producer/requirements.txt
pulsar-client==2.1.0
Faker==1.0.1
flake8==3.6.0
wheel==0.32.3